Add files via upload
This a simple Python code which completes a simple boolean function. The user is prompted to enter a value for "a" and a value for "b". The values are compared and either 1 or 0 is returned, based on whether the value of "a" is greater than or less than the value of "b".